services/events/public-service/requirements.txt
aiomysql
fastapi[standard]==0.115.0
fastapi-pagination==0.12.29
pydantic==2.9.2
pytz
pyjwt
aiomysql==0.2.0
sqlalchemy==2.0.36